Make the rationalize-dot macro play nicely with go
The `go` tool has a semantic meaning for the string "...", and I often need to
type it. Previously, the rationalize-dot macro would replace "..." with "../.."
to make it easier to work with relative paths. Now the macro replaces "...."
instead. This lets me type "..." and still save a keystroke on relative paths.